+++ mangooio-core/src/main/java/io/mangoo/annotations/Indexed.java
@@ -17,4 +17,5 @@ public @interface Indexed {
      */
     public Sort sort() default Sort.ASCENDING;
     public boolean unique() default false;
+    public boolean caseSensitive() default false;
 }